hw 24 getTotalInfoAboutEachFactory() editing
